refactor(viz): move edge_field_to_nodes into geode_util::viz
Relocate the FEM-viz edge-DOF → nodal field reconstruction out of the
example-only `geode-examples-support` crate and into the staging-layer
`geode_util::viz` module, so the Whitney reconstruction is a normal
library API rather than a sibling crate consumers depend on separately.

- Move `edge_field_to_nodes` plus its private Whitney/vector helpers
  (`cross`, `dot`, `sub`, `tet_grads`, `LOCAL_EDGES`) and the three unit
  tests into `crates/geode-util/src/viz.rs`, preserving the
  intentionally-crude per-tet-vertex averaging module docs. Behavior is
  identical (impl moved verbatim).
- Reduce `examples/_support` to a thin re-export
  (`pub use geode_util::viz::edge_field_to_nodes;`) so un-migrated
  consumers still compile; drop its `geode-core`/`faer` deps in favor of
  `geode-util`. Full decommission is tracked under Epic #414 Phase 3.
- Repoint the three consuming examples (mie_sphere, patch_antenna,
  spiral_inductor) to `geode_util::viz::edge_field_to_nodes` and drop
  their now-unused `geode-examples-support` dependency. Other functions
  in those crates (write_toml/sweep) are left untouched.
