+++ examples/_support/Cargo.toml
@@ -5,13 +5,13 @@ edition.workspace = true
 license.workspace = true
 repository.workspace = true
 rust-version.workspace = true
-description = "Shared, example-only viz glue for the GEODE-FEM standalone example crates."
+description = "Shared, example-only viz glue for the GEODE-FEM standalone example crates (thin re-export of geode_util::viz pending Phase 3 decommission)."
 
 [dependencies]
-geode-core = { workspace = true }
-# `c64` for the complex edge-DOF reconstruction; the workspace base faer
-# feature set (`std`, `linalg`) is sufficient — no `sparse-linalg` here.
-faer = { workspace = true }
+# The reconstruction now lives in `geode_util::viz`; this crate is a thin
+# re-export (Epic #414 Phase 2, issue #419) pending full decommission in
+# Phase 3.
+geode-util = { workspace = true }
 
 [lints.rust]
 unsafe_code = "forbid"
