+++ examples/patch_antenna/Cargo.toml
@@ -15,10 +15,9 @@ path = "src/main.rs"
 clap = { workspace = true }
 geode-app = { workspace = true }
 geode-core = { workspace = true, features = ["testing"] }
-geode-util = { workspace = true }
-# `edge_field_to_nodes` for the `--export-field` / `--export-sweep`
+# `viz::edge_field_to_nodes` for the `--export-field` / `--export-sweep`
 # Whitney edge-DOF → nodal E reconstruction.
-geode-examples-support = { workspace = true }
+geode-util = { workspace = true }
 # `BackendTypes` device handle for the driven solves.
 burn = { workspace = true }
 # `faer::c64` for the port impedance / S11 complex arithmetic.
